+++ library/src/actions/index.ts
@@ -1,3 +1,4 @@
+export * from './args/index.ts';
 export * from './await/index.ts';
 export * from './base64/index.ts';
 export * from './bic/index.ts';
@@ -71,6 +72,7 @@ export * from './rawTransform/index.ts';
 export * from './readonly/index.ts';
 export * from './reduceItems/index.ts';
 export * from './regex/index.ts';
+export * from './returns/index.ts';
 export * from './safeInteger/index.ts';
 export * from './size/index.ts';
 export * from './someItem/index.ts';