+++ library/CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to the library will be documented in this file.
 - Add `checkItemsAsync` action (pull request #856)
 - Add `graphemes`, `maxGraphemes`, `minGraphemes` and `notGraphemes` action (pull request #853)
 - Add `words`, `maxWords`, `minWords` and `notWords` action
+- Add `args` and `returns` action to transform functions (issue #243)
 - Change types and implementation to support Standard Schema
 - Change behaviour of `minValue` and `maxValue` for `NaN` (pull request #843)
 - Refactor `bytes`, `maxBytes`, `minBytes` and `notBytes` action